Refactor query service tests
Move the tests for various error conditions into a separate file
from the tests for successes, since the file was getting kind of
long. Move one helper function into `tests.support`.
